shuffled classes into files so I could call vat.ilk from an auction class
